design: (BH-873) Add base styles to settings page (#8)
* design: enqueue Open Sans from Google Fonts

* design: add HTML examples for main pages to app

These exist as placeholders for now. We can
comment them out or move them to a reference file
once we start building interactive components.

The HTML needs to be extracted into components
with the same structure and class names.

* design: add admin page heading markup

Also uses new page title from the Figma mockups.

* test: adjust test to look for new page title

* design: add styling for main views

A starting point. I expect this to change as
we build out the components.

* chore: fix chrome warnings about bad properties