refactor: pass 'text' instead of 'new' field type (#25)
Removes unnecessary use of useEffect in Field.jsx.

The 'open' prop now denotes a field that should be
open and treated as new or editable; the 'new'
type is not used.

Also prevents the flicker when first opening a
new field form. The form no longer needs to
re-render when useEffect changes the type from
'new' to 'text'.