chore: close field if no update occurred
Our REST endpoint returns a 200 status with
success:fail if all went well but update_option
returns false (no data was updated).

Closing the field in this case signifies that the
update was effectively successful and the user
needs to take no further action.

An alternative would be to present a message to
say “no changed were made”, but the user then has
to close the field manually by clicking the
form's “cancel” button.