BH-879: Add rich text field (#31)
* chore: create open and close field actions

Pass to fields along with editing state.

* chore: alter form appearance if editing

* chore: load existing data in name and slug fields

Loading data in additional fields to follow.

* chore: pass data to extra fields

* chore: default to stored data when editing fields

* chore: use id for openField for consistency

* chore: use editing form title when editing a field

* chore: improve formatting of component props

* chore: disable extra fields when editing

To meet acceptance criteria in BH-885.
https://wpengine.atlassian.net/browse/BH-885

- User may only edit field name and identifier.

* design: fade out disabled fields

To make it more clear visually that they are
not editable.

* fix: set initialState of nameCount to stored value

* chore: send PUT request to edit an existing field

Not strictly required but semantically more
correct than sending a POST request (which is
intended to create a new resource).

* chore: close field if no update occurred

Our REST endpoint returns a 200 status with
success:fail if all went well but update_option
returns false (no data was updated).

Closing the field in this case signifies that the
update was effectively successful and the user
needs to take no further action.

An alternative would be to present a message to
say “no changed were made”, but the user then has
to close the field manually by clicking the
form's “cancel” button.

* docs: improve fields rest route comment

* refactor: accept PUT requests to field endpoint

Not strictly required — we could just send a POST
request to update the field. But PUT is more
correct semantically.

This commit also updates the callback name to use
'update' in place of 'create', which seems more
accurate.

* fix: exclude edited field from slug exists check

So that changing a field value such as the title
(but leaving the API ID as-is, if we plan to allow
this later) does not cause the field to report an
error because its slug collides with itself.

* tests: adjust REST field exists check

A repeated request with the same ID would now
result in a 200 status instead of a 400 status to
allow for updates of that field.

This adjusts the test logic to check that a 400
status is only returned if two slugs with
different IDs are POSTed.

* tests: add test to check a field can be updated

* fix: name field label turns red only with errors

* design: add date field icon

* feat: add date to supported fields

* design: add Rich Text icon

* refactor: extract supported fields, add labels

Previously all field types were single words, so
the field labels could be inferred from the types
just be capitalizing the first letter.

“Rich Text” breaks that convention. We could
try to infer the label from a camel case
type name like "richText". But that will break
down if we ever have to translate labels in the
future.

This commit adds an object with field keys and
labels so that you can specify the label in plain
English. It could be extended to translate field
label names later if needed.

supportedFields.js is its own file because the
field label is used by Field.jsx and Form.jsx.

* fix: add default icon in case field type removed