wip: get fields data from models context
This commit breaks all field operations (add,
remove, open, delete, reorder) for the time being.

Also refactors to remove field 'order' from the
data. Instead of updating the order any time the
data changes, we can calculate field order when it
is needed. This will reduce complexity of the
field reducer functions so that they only have to
update the data and not the order.