Refactor data handling (#36)
* chore: add reducer file and model reducers

* refactor: useReducer instead of useState

To start storing state mutations in the same
place for both model and field operations.

* refactor: return model data from create REST route

Instead of requesting full model data to refresh
the model list when a model has been created, we
can return the new model from the REST endpoint
and update the model list in our new reducer.

This saves a request and updates the model list
more quickly.

* refactor: use reducer for model view and deletion

* refactor: use reducer for model creation

* chore: formatting

* refactor: show loading until data is loaded

* chore: formatting

* wip: get fields data from models context

This commit breaks all field operations (add,
remove, open, delete, reorder) for the time being.

Also refactors to remove field 'order' from the
data. Instead of updating the order any time the
data changes, we can calculate field order when it
is needed. This will reduce complexity of the
field reducer functions so that they only have to
update the data and not the order.

* refactor: use the reducer to open the field

* refactor: use the reducer to close the field

* refactor: use the reducer to update the field

* refactor: use the reducer to create a field

* fix: adjust reducer to work with empty fields

If a model has no fields yet this code fails
because the fields property is unset.

* refactor: use the reducer to remove a field

* refactor: remove unused addField function

* fix: remove stray semicolon

* refactor: move delete action to dropdown component

To avoid having to pass deleteAction down through
two components.

* refactor: swap field position via the reducer

* chore: formatting

* refactor: remove unused deleteAction prop

* refactor: extract field queries into own file

To keep these from accumulating in
EditContentModel, and to collect pure
functions that query model/field state without
fetching or sending data.

* refactor: load initial model state in a JS global

This will be present when the settings page loads,
so it saves us from having to do an initial fetch.

It should improve perceived load time even if
the initial server request for the settings page
takes a fraction longer.

* refactor: remove null checks

models can no longer be null now that it is being
populated from the initialState set by WP (which
gives an empty array if there are no models).

* chore: formatting

* fix: remove unused wpe/content-models GET endpoint

As discussed in https://github.com/wpengine/wpe-content-model/pull/36#discussion_r598925216.

* chore: remove unused prop