design: manually adjust button size
Had to crop canvas to 32px manually so that the
SVG would be centered in the button.

Figma generates SVGs with a large white space
around them and I couldn't find a way to adjust
this. It doesn't matter for most SVGs but did
here, where precise placement was important.