chore: add query to get root fields
I plan to store child fields flat with parent
fields, so that reducer functions can stay largely
the same and modify the field at the root level.
(If we store child fields in other fields, we have
to add messy, “if field has a parent, add/remove
it in the 'subfields' property of that parent
field, otherwise add/remove it from the 'fields'
property of the model itself” logic to the reducer
paths for add/open/close/etc. This also won't
scale well if repeater fields later have repeater
fields as children.)

The getRootFields function takes a flat list of
fields and gives a parent-child tree for display
to the user. It supports multiple parent-child
branches with a view of repeating repeater fields
one day.