BH-907: Add repeater field (#40)
* design: add repeater field icon

* feat: add repeater field

* design: add small icon

* chore: add empty repeater field state

* chore: add query to get root fields

I plan to store child fields flat with parent
fields, so that reducer functions can stay largely
the same and modify the field at the root level.
(If we store child fields in other fields, we have
to add messy, “if field has a parent, add/remove
it in the 'subfields' property of that parent
field, otherwise add/remove it from the 'fields'
property of the model itself” logic to the reducer
paths for add/open/close/etc. This also won't
scale well if repeater fields later have repeater
fields as children.)

The getRootFields function takes a flat list of
fields and gives a parent-child tree for display
to the user. It supports multiple parent-child
branches with a view of repeating repeater fields
one day.

* fix: count and display only root fields

Subfields of repeater fields should not be counted
towards the field total or displayed at the top
level of the field list.

* chore: create new fields with parent ID if passed

* feat: add ability to add a subfield

TODO: save parent id when a subfield is created.

* chore: pass parent to form to store on submission

* chore: send full model data to repeater

* chore: fix deletion of repeater fields

Deleting a repeater field now deletes its
descendents, including nested repeaters in
subfields, so that no orphaned fields remain.

More complex deletion logic is the main downside
of storing fields in a flat list, but it seems
worthwhile because other field operations (adding,
changing 'open' state etc.) do not require
alterations to account for repeater fields.

* chore: formatting

* design: fix subfield button size and field spacing

* design: fix alignment of nested subfield + buttons

* fix: support subfield reordering

By re-using updatePositions logic from the parent
view, adapted to send subfield order when any
fields are reordered for that model, instead of
just root field order.

This could be improved by only sending the
new positions of fields that have changed order
(instead of all fields). It would mean tracking
clean/dirty state of field data and would make a
good future improvement to reduce the size of
the REST request payload with very large
fieldsets.

Subfields of a repeater start at position 0. They
are ordered independently of parent fields so that
their position is not affected by the parent
repeater field's position.

Subfields can only be ordered at their current
level of nesting. You cannot order a field out of
its parent field, because that felt like a recipe
for pain and confusion.