[BH-922] Fix API ID field to generate correct IDs from field title and model plural name (#48)
* fix: add toValidApiId for legal API IDs

* chore: add custom hook to link name and API fields

* chore: amend forms to use API ID generator

* chore: remove camelcase dependency

* chore: bump maxLength of API field to 50

To match the input field max. 50 is still an
arbitrary limit, but this will reduce the chance
of forms being rejected when an API ID has been
generated from a title exceeds 20 characters.

* chore: lower case initial capitals

So that IDs start with a lowercase letter. Not
required by the GraphQL spec but consistent with
common usage.

* test: check leading underscores are permitted

* test: check accented characters are stripped

* refactor: remove trimStart call

This scenario is now covered by making the initial
letter lowercase.

* chore: remove old comment

* fix: revert maxLength for post slug to 20

There's a hard-limit in register_post_type:

if ( empty( $post_type ) || strlen( $post_type ) > 20 ) {
    _doing_it_wrong( __FUNCTION__, __( 'Post type names must be between 1 and 20 characters in length.' ), '4.2.0' );
    return new WP_Error( 'post_type_length_invalid', __( 'Post type names must be between 1 and 20 characters in length.' ) );
}

Props @mindctrl.