feat: show success toast after models are created
In response to feedback from customers and design
amends from Mikolaj and Mason.

We found that customers were not sure that their
model had been created because it happens so fast,
and because we move to the model edit screen
instead of the model list screen when a model
has been created.