chore: pass models to onDragEnd
So that onDragEnd can restore the original field
order if the position update fails. Otherwise the
user is left in a state where the field position
looks correct on screen even though it failed.

We prompt them to retry field ordering, but they
can't if the local state is incorrect without
dragging fields to the original order manually
and retrying.