chore: add 'submitted' class to the submitted form
Allows styling of required fields only when the
form has been submitted.

The HTML5 :invalid pseudoselector is applied to
a 'required' field as soon as the form loads.

Without this change, CSS targeting input:invalid
applies to new forms, so every required field is
styled with a red border even before the user has
entered any data.

Adding a class to the form on submission ensures
fields are only highlighted on form submission.