fix: prevent intermittent “field is missing” error
Opening Advanced Settings in a Text field and
clicking “Done” would occasionally result in a
“field is missing” error for the fields in the
modal, which prevented it from closing.

This commit addresses this by:
- Making advancedSettings a dep of useEffect,
to prevent possible timing issues where useEffect
could run when advancedSettings is undefined,
which prevented fields from being registered
(hence “field is missing”).
- Passing the value of the advanced fields into
the form at the top level as defaultValues instead
of using setValue. This prevents an issue caused
by the addition of advancedSettings as a dep where
values would be reset when closing the modal with
“Done” and then opening it again.