fix: amend check for at least one model
Sites with only one ACM model will present a
single checkbox in the list of models.

This is submitted as a string, not an array, so
we need to cast strings to arrays for consistent
storage.

Unchecked models are sent either as boolean
false (only one model) or an empty array (multiple
models), so this also amends the check for those
cases.