fix: relink fields on successful form submission
When someone edits the taxonomy API Identifier
manually, this breaks the “link” between the
Singular name field and the API Identifier field
(so that edits to the singular name no longer
update the API ID field, because this would
destroy any manual changes they made).

When someone edits the API ID field manually
(breaking the link) and then saves their new
taxonomy successfully, this link between the
fields needs to be reset so the fields are linked
again for the next taxonomy to be entered.

This ensures that the next taxonomy someone adds
will generate an API ID from the Singular field
again.

We did not have this issue with other forms,
because the form for models/fields does not remain
in view after submission.