fix: hide private taxonomies in REST responses
So that private taxonomies and their terms are not
available without authentication.

Previously, terms of private taxonomies would be
hidden but taxonomies would still be visible.

Props to @apmatthews for the discovery at
https://github.com/wpengine/atlas-content-modeler/pull/180#pullrequestreview-695111288.

The same issue does not affect WPGraphQL; our
existing filter is blocking both terms and
taxonomies for private taxonomies as designed.