fix: reduce private taxonomy capability to read
As discussed with @mindctrl and @apmatthews,
this reduces access requirements to read a private
taxonomy via the REST API to be more consistent
with our post type API (which checks that a user
can read the post).

A 'read' capability is the lowest requirement,
so this effectively changes access levels from
administrator to subscriber.

See https://wordpress.org/support/article/roles-and-capabilities/#capability-vs-role-table.