refactor: adjust state handling
To prevent direct mutation of state via props.

- Lifts selectedEntries state to parent so it can be set from
the modal.

- Introduces chosenEntries in modal as a way to track changes
of the chosen entries.

- Updates chosen entries when the user makes selections. (This is
needed because selections may appear across multiple pages, only
one of which is rendered at a time — we cannot just map over checkbox
inputs when the user saves the field because not all of those inputs
will still be rendered.)

- Persists those changes to the parent via setSelectedEntries when the
user saves changes.

- Switches to .filter instead of .splice to remove checked items.
(.splice mutates the state value in place, which meant unticking an item
would update the list of linked entries immediately instead of when
changes are saved).

- Renames functions and variables to use consistent terminology
(“entries” instead of a mix of “entries” and “related”).