refactor: allow all entries to be removed
So that a multi relationship field's choices can all be removed.

Before this, you could untick all boxes but the originally chosen
entries would still appear, because unticking fired a request with
an empty query (no IDs are selected). This meant WP responded
with a list of all entries of that type, which would appear when
you clicked Save in the modal.