chore: initialize analytics in the settings app
We can't send the pageview at the App root level here, because every
view uses this App as the wrapper, so a visit to the taxonomies page
will trigger “ACM Models Home” incorrectly.

We can trigger the pageview from the specific view instead.

This should eventually be gated by a check for user opt-in.