test: just check the modal is still open
Unable to check if the modal refreshes due to Circle running
Chrome in headless mode, where it does not fire the
visibilitychange events that trigger the modal refresh.

I was able to confirm this by changing acceptance.suite.yml
to use `args: ["disable-gpu", "headless"]` locally to get it
to behave the same as the acceptance.suite.circleci.yml.