fix: mount publisher app even if no fields present
Required because the “Empty Trash” modal needs to load even in the
absence of fields. The “Move to Trash” link is still available in the
publish meta box on entries with missing fields, and we need to be
able to intercept clicks on the move to trash link.

I added a “no fields found” message instead of a blank screen.