[BH-1140] Show warning when trashing an entry that is referenced in a relationship field (#263)
* chore: test if publisher entry has references

So that we can intercept “Send to Trash” clicks only if a post
has references.

An alternative to performing this check up front is to do it
when the trash link is clicked. This depends on the roundtrip
speed, though, and may feel laggy on slower connections.

* wip: intercept trash link clicks

Proof of concept. Using `confirm()` for now but intend to swap this
with a React Modal.

* Revert "wip: intercept trash link clicks"

This reverts commit 78682297044849fd61501475674c616a012aabe2.

* feat: add trash post confirmation modal

This appears in the Publisher screen (not on the post listing index).

* chore: remove unused code

* chore: simplify modal language

* chore: use model name in trash modal title

* chore: trim word from explanation

* fix: use string type as post ID

int causes a fatal error.

* fix: mount publisher app even if no fields present

Required because the “Empty Trash” modal needs to load even in the
absence of fields. The “Move to Trash” link is still available in the
publish meta box on entries with missing fields, and we need to be
able to intercept clicks on the move to trash link.

I added a “no fields found” message instead of a blank screen.

* test: add e2e test for trash confirmation modal

* fix: add wait(1) for flaky test

This test has failed twice on different branches.

Screenshots show the text does not appear in the menu item.

Trying a wait(1) in case we're checking too quickly before the
re-render has occurred.

* test: wait for relationship entry to open