fix: limit field count check to Gutenberg
I previously changed enqueue_assets to
prevent scripts loading if field count is less
than 1 instead of if the current post type is
not an ACM model.

This was needed so that the scripts can load
on non ACM post types where relationship
fields are now displayed, such as posts and
pages.

But the app still needs to load on the ACM
publisher entry pages because we display
a “No fields were found for this entry's
model.” message there when there are no
fields (and have tests to check for this).