refactor: extract post_type defaults
Pass all post types to collect_posts instead
of specifying the defaults in the function.

Allows us to pass post_types to other collection
functions, so that collect_post_tags and others
can act based on the post types the
user has requested.

Note that this does not change the public-facing
API (the user still has to specify overrides,
and the defaults are the same), so there is
no change of usage.