chore: add 'String' case to switch statement.
If a field's type has changed from the default internal type
to a type to be registered with WPGraphQL, running it through
this switch again results in the `default` case running where
it is not desirable.
e.g. a `text` field type is converted to `String` before
being registered with WPGraphQL.
See `map_html_field_type_to_graphql_field_type()` for more info
on field types.