MTKA-1302: Add Rich Text repeatable option (#462)
Rich Text fields now include an optional 
“Make this field repeatable” checkbox.

It behaves the same way as the Text field's 
repeatable option, allowing publishers to input 
multiple rich text fields as separate rows.

Adds the uuid JS dependency to generate unique IDs
for repeatable rich text fields.

We can't use the index of repeatable rich text
fields as their ID, because in a repeatable context
this is not a unique value (an item with index of '3' has
its ID changed to '2' if you delete the item at index '2'
by clicking the “delete item” button).

These IDs need to stay constant and stable regardless
of their position in the list. So we give each a unique ID
for the purpose of initializing TinyMCE wherever they
appear.

The unique ID also doubles as the key for use in React.
This prevents unneeded re-renders when deleting items,
and stops quirks such as deleting an item from the middle
of the list causing the one at the end of the list to be removed.